Move image classes into ndimage.h
